Fix errors discovered by static analysis

+++ inc/template-tags.php
@@ -170,7 +170,7 @@ function _s_posted_on() {
 		esc_attr( get_the_time() ),
 		esc_attr( get_the_date( 'c' ) ),
 		esc_html( get_the_date() ),
-		esc_url( get_author_posts_url( get_the_author_meta( 'ID' ) ) ),
+		esc_url( get_author_posts_url( (int) get_the_author_meta( 'ID' ) ) ),
 		esc_attr( sprintf( __( 'View all posts by %s', '_s' ), get_the_author() ) ),
 		esc_html( get_the_author() )
 	);
